Delete `ConsentStatusParamRequest` class
