+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/data/network/model/optimized/сonsentStatus/GranularState.kt
@@ -1,0 +1,12 @@
+package com.sourcepoint.cmplibrary.data.network.model.optimized.сonsentStatus
+
+enum class GranularState {
+    ALL,
+    SOME,
+    NONE,
+    EMPTY_VL;
+
+    companion object {
+        fun firstWithStateOrNONE(state: String?) = (entries.firstOrNull { it.name == state })?: NONE
+    }
+}
